backend/requirements.txt
fastapi
uvicorn
langchain
langchain_huggingface
python-dotenv
pydantic